# docs/swagger.yaml
definitions:
  types.ApiObjDetail:
    properties:
      action:
        type: string
      createBy:
        type: integer
      createdAt:
        type: string
      handle:
        type: string
      id:
        description: convert to string id
        type: string
      path:
        type: string
      title:
        type: string
      type:
        type: string
      updateBy:
        type: integer
      updatedAt:
        type: string
    type: object
  types.Column:
    properties:
      exp:
        description: expressions, which default to = when the value is null, have
          =, !=, >, >=, <, <=, like
        type: string
      logic:
        description: logical type, defaults to and when value is null, only &(and),
          ||(or)
        type: string
      name:
        description: column name
        type: string
      value:
        description: column value
    type: object
  types.Conditions:
    properties:
      columns:
        description: columns info
        items:
          $ref: '#/definitions/types.Column'
        type: array
    type: object
...
  types.DeleteApisByIDsRequest:
    properties:
      ids:
        description: id list
        items:
          type: integer
        minItems: 1
        type: array
    type: object
  types.DeleteApisByIDsRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
      msg:
        description: return information description
        type: string
    type: object
...
  types.GetApiByConditionRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          api:
            $ref: '#/definitions/types.ApiObjDetail'
        type: object
      msg:
        description: return information description
        type: string
    type: object
...
  types.ListApisByIDsRequest:
    properties:
      ids:
        description: id list
        items:
          type: integer
        minItems: 1
        type: array
    type: object
  types.ListApisByIDsRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          apis:
            items:
              $ref: '#/definitions/types.ApiObjDetail'
            type: array
        type: object
      msg:
        description: return information description
        type: string
    type: object
  types.ListApisRespond:
    properties:
      code:
        description: return code
        type: integer
      data:
        description: return data
        properties:
          apis:
            items:
              $ref: '#/definitions/types.ApiObjDetail'
            type: array
        type: object
      msg:
        description: return information description
        type: string
    type: object
...
  types.Params:
    properties:
      columns:
        description: query conditions
        items:
          $ref: '#/definitions/types.Column'
        type: array
      page:
        description: page number, starting from page 0
        type: integer
      size:
        description: lines per page
        type: integer
      sort:
        description: sorted fields, multi-column sorting separated by commas
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: http server api docs
  title: admin api docs
  version: "2.0"
paths:
...
  /api/v1/api/condition:
    post:
      consumes:
      - application/json
      description: get api by condition
      parameters:
      - description: query condition
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.Conditions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.GetApiByConditionRespond'
      security:
      - BearerAuth: []
      summary: get api by condition
      tags:
      - api
  /api/v1/api/delete/ids:
    post:
      consumes:
      - application/json
      description: delete apis by batch id
      parameters:
      - description: id array
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.DeleteApisByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.DeleteApisByIDsRespond'
      security:
      - BearerAuth: []
      summary: delete apis
      tags:
      - api
  /api/v1/api/list:
    get:
      consumes:
      - application/json
      description: list of apis by last id and limit
      parameters:
      - default: 0
        description: last id, default is MaxInt32
        in: query
        name: lastID
        required: true
        type: integer
      - default: 10
        description: size in each page
        in: query
        name: limit
        type: integer
      - default: -id
        description: 'sort by column name of table, and the '
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListApisRespond'
      security:
      - BearerAuth: []
      summary: list of apis by last id and limit
      tags:
      - api
    post:
      consumes:
      - application/json
      description: list of apis by paging and conditions
      parameters:
      - description: query parameters
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.Params'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListApisRespond'
      security:
      - BearerAuth: []
      summary: list of apis by query parameters
      tags:
      - api
  /api/v1/api/list/ids:
    post:
      consumes:
      - application/json
      description: list of apis by batch id
      parameters:
      - description: id array
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/types.ListApisByIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListApisByIDsRespond'
      security:
      - BearerAuth: []
      summary: list of apis by batch id
      tags:
      - api
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type Bearer your-jwt-token to Value
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"